Split test_data and test_utils module + some formatting
