+++ src/intervals/absolute/half_bounded_interval_tests.rs
@@ -32,7 +32,7 @@ use crate::intervals::meta::{
     Relativity,
 };
 use crate::intervals::special::{EmptyInterval, UnboundedInterval};
-use crate::test_data::date_timestamp;
+use crate::test_utils::date_timestamp;
 
 mod new {
     use super::*;
