+++ src/intervals/ops/bound_cmp/bound_ord_tests.rs
@@ -3,14 +3,8 @@ use std::fmt::Debug;
 
 use crate::intervals::absolute::{AbsEndBound, AbsFiniteEndBound, AbsFiniteStartBound, AbsStartBound};
 use crate::intervals::meta::BoundInclusivity;
-use crate::intervals::ops::{
-    BoundOrd,
-    BoundOrdering,
-    BoundOverlapAmbiguity,
-    BoundOverlapDisambiguationRuleSet,
-};
+use crate::intervals::ops::{BoundOrd, BoundOrdering, BoundOverlapAmbiguity, BoundOverlapDisambiguationRuleSet};
 use crate::intervals::relative::{RelEndBound, RelFiniteEndBound, RelFiniteStartBound, RelStartBound};
-use crate::test_data::BinOpPair;
 use crate::test_data::bound_overlap::{
     FINITE_END_FINITE_END_ABS,
     FINITE_END_FINITE_END_REL,
@@ -45,6 +39,7 @@ use crate::test_data::bound_overlap::{
     INF_START_INF_START_ABS,
     INF_START_INF_START_REL,
 };
+use crate::test_utils::BinOpPair;
 
 type BodRuleSet = BoundOverlapDisambiguationRuleSet;
