+++ src/lib.rs
@@ -67,6 +67,8 @@ pub mod ops;
 pub mod prelude;
 pub mod time;
 
+mod test_utils;
+
 #[cfg(feature = "arbitrary")]
 mod arbitrary_impl;
